+++ modules/tinymce/src/core/main/ts/api/ui/Ui.ts
@@ -17,9 +17,8 @@ type Registry = BridgeRegistry.Registry;
 export interface EditorUiApi {
   show: () => void;
   hide: () => void;
-  enable: () => void;
-  disable: () => void;
-  isDisabled: () => boolean;
+  setEnabled: (state: boolean) => void;
+  isEnabled: () => boolean;
 }
 
 export interface EditorUi extends EditorUiApi {