+++ modules/tinymce/src/themes/silver/test/ts/headless/window/CustomDialogTest.ts
@@ -119,7 +119,7 @@ describe('headless.tinymce.themes.silver.window.CustomDialogTest', () => {
         {
           type: 'custom',
           text: 'go',
-          disabled: true
+          enabled: false,
         },
         {
           type: 'cancel',